requirements.txt
requests~=2.31.0
urllib3~=2.2.1
fastapi~=0.110.0
python-dotenv~=1.0.1
pydantic~=2.6.4
uvicorn~=0.29.0
